+++ test/Microsoft.ServiceFabric.Actors.IntegrationTests/TestableActorService.cs
@@ -1,0 +1,37 @@
+using Fuzzy;
+using Microsoft.ServiceFabric.Actors.Runtime;
+using Microsoft.ServiceFabric.Services.Runtime;
+using System;
+using System.Fabric;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.ServiceFabric.Actors
+{
+    public class TestableActorService
+    {
+        static readonly IFuzz fuzzy = new RandomFuzz();
+
+        public static async Task<ActorService> GetActorService<T>(
+            Func<ActorService, ActorId, ActorBase> actorFactory = null,
+            ActorServiceSettings actorServiceSettings = null,
+            IActorStateProvider actorStateProvider = null)
+            where T : Actor
+        {
+            ActorService actorService = new ActorService(
+                fuzzy.StatefulServiceContext(),
+                ActorTypeInformation.Get(typeof(T)),
+                actorFactory,
+                null,
+                actorStateProvider ?? new NullActorStateProvider(),
+                actorServiceSettings);
+
+            IStatefulUserServiceReplica statefulServiceReplica = actorService;
+            await statefulServiceReplica.OnOpenAsync(ReplicaOpenMode.New, CancellationToken.None);
+            await statefulServiceReplica.OnChangeRoleAsync(ReplicaRole.Primary, CancellationToken.None);
+            await statefulServiceReplica.RunAsync(CancellationToken.None);
+
+            return actorService;
+        }
+    }
+}
